Refactor form register
