+++ src/App.tsx
@@ -17,10 +17,6 @@ import ViewPost from './pages/ViewPost/ViewPost';
 import Register from './pages/Register/Register';
 import CreatePost from "./pages/CreatePost/CreatePost";
 import MarketPlace from "./pages/MarketPlace/MarketPlace";
-import FormPersonas from "./pages/Register/FormPersonas";
-import FormEmpresas from "./pages/Register/FormEmpresas"
-
-import { firebaseConfig } from "./config/environtment";
 
 /* Core CSS required for Ionic components to work properly */
 import "@ionic/react/css/core.css";
@@ -53,8 +49,6 @@ const Router = (
         <Route exact path="/register" component={Register} />
         <Route path="/marketplace" component={MarketPlace} />
         <Route exact path="/" component={Home} />
-        <Route exact path="/register/personas" component={FormPersonas} />
-        <Route exact path="/register/empresas" component={FormEmpresas} />
       </IonRouterOutlet>
     </IonSplitPane>
   </IonReactRouter>
