--- src/pages/Register/FormPersonas.css
@@ -1,7 +1,0 @@
-.swiper-slide{
-  display: block;
-}
-
-ion-slide{
-  height: 100%;
-}
